Latex feature pages: Enable DND-tables on fancy sheets
Dungeonsheets uses rst-to-latex to generate latex text. For tables, rst-to-latex
uses the latex supertabular format. However, the RPGtex latex style's dndtables
are based on a different table formate, that is, tabularx, which is not entirely
compatible with supertabular:
  - Table headings result in latex errors
  - Columns end up too wide because text wrapping is lacking.

This patch modifies the rst-to-latex output so that it generates a correct table
header for the dndtable format (if a header is present), uses the X column
format in dndtables to allo for text wrapping, and it translates the column
widths as computed by rst-to-latex for the supertabular format to something
that dndtable can use.

The resulting tables are all slightly more compact than the regular supertabular
tables, and they look better :-)

This patch only uses dndtables for the fancy sheets; regular feature pages
keep using the supertabular output in order for them not to depend on
the RPGtex latex style.